Changed % to modulo. And for that purpose changed M to long int so that it doesnt overflow to negative when casting it from (int) to (long int). Created Hypercube in a different folder (main differences: one hashtable in main and in search, euclidean does not require dimension reduce but just an modulo2 for every hi in H.

+++ HypercubeLSH/include/Euclidean.hpp
@@ -13,7 +13,7 @@ class Euclidean: public Metric{
     std::vector<float> t; //K single precision random reals
     static int w;
     static std::vector<int> r; //K random integers<=32bit
-    static long int M;
+    static unsigned int M;
   public:
     Euclidean(int dim, int tsize);
     ~Euclidean();
